open library

+++ src/main/java/com/api/gesboo/repository/BookRepository.java
@@ -5,9 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
     Book findByIsbn(String isbn);
-
-//    List<Book> findAll(List<Book> books);
 
 }
